Spawning of planes

+++ main/java/minecrafttransportsimulator/systems/pack/PackDisplayTextConfig.java
@@ -4,6 +4,6 @@ package minecrafttransportsimulator.systems.pack;
 public class PackDisplayTextConfig {
 
     public float[] pos;
-    public int[] rot;
+    public float[] rot;
 
 }
