+++ main/java/minecrafttransportsimulator/systems/pack/PackInstrument.java
@@ -4,7 +4,7 @@ package minecrafttransportsimulator.systems.pack;
 public class PackInstrument {
 
     public float[] pos;
-    public int[] rot;
+    public float[] rot;
     public int defaultInstrument;
 
 }
