+++ main/java/minecrafttransportsimulator/systems/pack/PackPart.java
@@ -4,7 +4,7 @@ package minecrafttransportsimulator.systems.pack;
 public class PackPart {
 
     
-    public int[] pos;
+    public float[] pos;
     public boolean turnsWithSteer;
     public boolean isController;
     public String[] names;
